docker: add raspi support

+++ docker.sh
@@ -1,23 +1,24 @@
-sunzi.mute apt-get remove --yes --purge docker docker-engine docker.io containerd runc
-sunzi.install apt-transport-https \
-              ca-certificates \
-              curl \
-              gnupg-agent \
-              software-properties-common
+main() {
+  arch=$(uname -m)
+  march=$(echo $arch | sed s/armv7l/armhf/ | sed s/x86_64/amd64/)
+  if [ "$arch" != x86_64 ] && [ "$arch" != armv7l ]; then
+   echo "syncthing: architecture not supported" >&2
+   return
+  fi
+  sunzi.mute apt-get remove --yes --purge docker docker-engine docker.io containerd runc
+  sunzi.install apt-transport-https \
+                ca-certificates \
+                curl \
+                gnupg-agent \
+                software-properties-common
 
-curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
+  curl -sSL https://get.docker.com | sh
 
-sunzi.mute add-apt-repository --yes \
-           "deb [arch=amd64] https://download.docker.com/linux/debian \
-           $(lsb_release -cs) \
-           stable"
+  sunzi.mute curl -s -L "https://github.com/docker/compose/releases/download/1.25.5/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+  sunzi.mute chmod +x /usr/local/bin/docker-compose
+  if [ -n "$SUDO_USER" ]; then
+    sunzi.mute gpasswd -a $SUDO_USER docker
+  fi
+}
 
-sunzi.mute apt-get update
-
-sunzi.install docker-ce docker-ce-cli containerd.io
-
-sunzi.mute curl -s -L "https://github.com/docker/compose/releases/download/1.25.5/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-sunzi.mute chmod +x /usr/local/bin/docker-compose
-if [ -n "$SUDO_USER" ]; then
-  sunzi.mute gpasswd -a $SUDO_USER docker
-fi
+main
